requirements.txt
flask
transformers
torch
faiss-cpu
sentence-transformers